+++ changelog.rst
@@ -7,6 +7,8 @@ ChangeLog
 
 .. note:: This version is not yet released and is under active development
 
+* Load up all subfolders from ``Maildir`` and ``MH`` boxes. Closes #123.
+
 
 `6.0.1 (2020-10-22) <https://github.com/kdeldycke/mail-deduplicate/compare/v6.0.0...v6.0.1>`_
 ---------------------------------------------------------------------------------------------